DONTPUSH examples/iscsi-dd: switch src and dst CSCD desc ids
Reproducer for:
https://bugzilla.kernel.org/show_bug.cgi?id=191381

The source and destination CSCD descriptor IDs in the segment descriptor
should be used to determine the corresponding copy source and
destination devices.
LIO incorrectly assumes that CSCD descriptors are listed in
source-then-dest order, so switching the order (and the corresponding
CSCD source and destination descriptor IDS) causes LIO to use the wrong
source and target devices.

Tagged DONTPUSH, as we don't really want this behaviour in a dd utility.